requirements.txt
ultralytics
opencv-python
pytesseract
streamlit
matplotlib
Pillow
torch>=2.2
torchvision
roboflow
fastapi
uvicorn
rapidfuzz
numpy
tensorflow
scikit-learn